adjust tests so lastLogin is tested for both implementations
